Move to tickets.js
