+++ src/manager/ResourceManager.js
@@ -1,7 +1,4 @@
 import path from 'path';
-import { config } from '../index.js';
-
-import fetch from 'node-fetch';
 
 export default class ResourceManager {
   constructor() {
@@ -20,45 +17,6 @@ export default class ResourceManager {
 
       // use hostname/filename as identifier:
       const processedURL = new URL(url);
-
-      const safeBrowsingURL =
-        'https://safebrowsing.googleapis.com/v4/threatMatches:find?key=' +
-        index.config.google_safe_browsing_api_key;
-      const request = {
-        client: {
-          clientId: '4950',
-          clientVersion: '0.0.1',
-        },
-        threatInfo: {
-          threatTypes: ['MALWARE'],
-          platformTypes: ['ANY_PLATFORM'],
-          threatEntryTypes: ['URL'],
-          threatEntries: [{ url: processedURL }],
-        },
-      };
-
-      try {
-        fetch(safeBrowsingURL, {
-          method: 'POST',
-          body: JSON.stringify(request),
-          responseType: 'application/json',
-        })
-          .then(res => res.json())
-          .then(res => {
-            if (res.matches === undefined) {
-              // console.log('----------------------------------------------');
-              // console.log('no malware find');
-            } else {
-              const { matches } = res.matches;
-              const match = matches[0];
-              console.log('----------------------------------------------');
-              console.log('malware find:');
-              console.log(match);
-            }
-          });
-      } catch (error) {
-        console.log(error);
-      }
 
       const hostname = processedURL.hostname;
       // grab the filename from the path specifier:
