+++ src/routes/ticket.js
@@ -19,13 +19,13 @@ router.post('/', async (req, res) => {
     );
 
     const ticket = new Ticket(ID, url, screenshots);
-    const artifacts = await ticketManager.processTicket(ticket);
+    const results = await ticketManager.processTicket(ticket);
 
     console.log(
       `Finished processing ticket with ID ${ticket.getID()}! Awaiting more...`
     );
 
-    res.json(artifacts);
+    res.json(results);
   } catch (e) {
     console.error('An error occurred when processing this ticket.');
     console.log(e);
